requirements.txt
pywebview>=5.0
zstandard>=0.20.0
